backend/requirements.txt
# LocalBrisk Backend 依赖
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
polars>=0.20.0
duckdb>=0.9.0
httpx>=0.26.0
pyyaml>=6.0.0

# Agent 引擎依赖
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
deepagents>=0.4.3

# 数据库驱动
pymysql>=1.1.0
cryptography>=41.0.0

# 打包工具
pyinstaller>=6.0.0

# 测试依赖
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
